+++ projects/core/src/main/kotlin/site/siredvin/turtlematic/computercraft/plugins/AutomataInteractionPlugin.kt
@@ -7,11 +7,13 @@ import dan200.computercraft.api.lua.MethodResult
 import net.minecraft.world.entity.Entity
 import net.minecraft.world.item.ItemStack
 import site.siredvin.peripheralium.api.datatypes.InteractionMode
-import site.siredvin.peripheralium.api.datatypes.VerticalDirection
 import site.siredvin.peripheralium.api.peripheral.IPeripheralOperation
 import site.siredvin.peripheralium.computercraft.peripheral.owner.TurtlePeripheralOwner
 import site.siredvin.turtlematic.computercraft.operations.SingleOperation
 import site.siredvin.turtlematic.computercraft.peripheral.automatas.BaseAutomataCorePeripheral
+import site.siredvin.turtlematic.util.getInteractionMode
+import site.siredvin.turtlematic.util.optPose
+import site.siredvin.turtlematic.util.optVerticalDirection
 import java.util.function.Predicate
 
 class AutomataInteractionPlugin(
@@ -23,21 +25,18 @@ class AutomataInteractionPlugin(
         get() = listOf(SingleOperation.SWING, SingleOperation.USE)
 
     fun swingImplInner(arguments: IArguments): MethodResult {
-        val mode = InteractionMode.luaValueOf(arguments.getString(0), allowedMods)
-        val directionArgument = arguments.optString(1)
-        val overwrittenDirection = if (directionArgument.isEmpty) {
-            null
-        } else {
-            VerticalDirection.luaValueOf(
-                directionArgument.get(),
-            )
-        }
+        val mode = arguments.getInteractionMode(0, allowedMods)
+        val overwrittenDirection = arguments.optVerticalDirection(1)
+        val overwrittenPose = arguments.optPose(2)
         if (!allowedMods.contains(mode)) {
             return MethodResult.of(null, "Mode $mode are not allowed for this core")
         }
         automataCore.addRotationCycle()
         val owner = automataCore.peripheralOwner
         val result = owner.withPlayer({
+            if (overwrittenPose != null) {
+                it.fakePlayer.pose = overwrittenPose
+            }
             it.swing(skipEntity = mode.skipEntry, skipBlock = mode.skipBlock, entityFilter = suitableEntity)
         }, overwrittenDirection = overwrittenDirection?.minecraftDirection)
         if (!result.left) {
@@ -60,21 +59,18 @@ class AutomataInteractionPlugin(
     }
 
     fun useImplInner(arguments: IArguments): MethodResult {
-        val mode = InteractionMode.luaValueOf(arguments.getString(0), allowedMods)
-        val directionArgument = arguments.optString(1)
-        val overwrittenDirection = if (directionArgument.isEmpty) {
-            null
-        } else {
-            VerticalDirection.luaValueOf(
-                directionArgument.get(),
-            )
-        }
+        val mode = arguments.getInteractionMode(0, allowedMods)
+        val overwrittenDirection = arguments.optVerticalDirection(1)
+        val overwrittenPose = arguments.optPose(2)
         if (!allowedMods.contains(mode)) {
             return MethodResult.of(null, "Mode $mode are not allowed for this core")
         }
         automataCore.addRotationCycle()
         val owner = automataCore.peripheralOwner
         val result = owner.withPlayer({
+            if (overwrittenPose != null) {
+                it.fakePlayer.pose = overwrittenPose
+            }
             it.use(skipEntity = mode.skipEntry, skipBlock = mode.skipBlock, entityFilter = suitableEntity)
         }, overwrittenDirection = overwrittenDirection?.minecraftDirection)
         return MethodResult.of(true, result.toString())