+++ projects/core/src/main/kotlin/site/siredvin/turtlematic/util/Extensions.kt
@@ -1,9 +1,16 @@
 package site.siredvin.turtlematic.util
 
+import dan200.computercraft.api.lua.IArguments
+import dan200.computercraft.api.lua.LuaException
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.util.Mth
+import net.minecraft.world.entity.Pose
 import net.minecraft.world.entity.projectile.Projectile
 import net.minecraft.world.phys.Vec3
+import site.siredvin.peripheralium.api.datatypes.InteractionMode
+import site.siredvin.peripheralium.api.datatypes.VerticalDirection
+import java.util.*
+import java.util.stream.Collectors
 
 fun ResourceLocation.toNetherite(): ResourceLocation {
     return ResourceLocation(this.namespace, "netherite_${this.path}")
@@ -29,4 +36,35 @@ fun Projectile.advancedShoot(deltaMovements: Vec3) {
     xRot = (Mth.atan2(deltaMovements.y, horizontalPower) * 57.2957763671875).toFloat()
     yRotO = yRot
     xRotO = xRot
+}
+
+fun IArguments.getInteractionMode(index: Int, allowedMods: Set<InteractionMode>): InteractionMode {
+    return InteractionMode.luaValueOf(this.getString(index), allowedMods)
+}
+
+fun IArguments.optVerticalDirection(index: Int): VerticalDirection? {
+    val directionArgument = this.optString(index)
+    return if (directionArgument.isEmpty) {
+        null
+    } else {
+        VerticalDirection.luaValueOf(
+            directionArgument.get(),
+        )
+    }
+}
+
+fun IArguments.optPose(index: Int): Pose? {
+    val poseArgument = this.optString(index)
+    return if (poseArgument.isEmpty) {
+        null
+    } else {
+        try {
+            Pose.valueOf(poseArgument.get().uppercase())
+        } catch (exc: IllegalArgumentException) {
+            val allValues = Arrays.stream(Pose.values()).map { mode -> mode.name.lowercase() }.collect(
+                Collectors.toList(),
+            ).joinToString(", ")
+            throw LuaException("Vertical direction should be one of: $allValues")
+        }
+    }
 }